added model Hamiltonian to Hilbert
In order for model Hamiltonian (Pauli-Fierz) to match our CQED-CIS, we need to have access to the ground-state and first-excited state total dipole moments, as well as the transition dipole moments!  Working on making a helper_cis.py that will compute the total dipole moments for the ground- and excited-states, since the latter are not available in psi4 to my knowledge!